Move `RemoteCEPGetterMapper` to its own file
